Moved auto assign button into its own component - moved to organization queue (#20743)
